ABI future-proofing: pin enum width, fixed-width uint64_t sizes (#106)
* fix(abi): pin enum width and use uint64_t in the C ABI (future-proofing)

Two layout-neutral hardening changes to the plugin C ABI, both protecting
third-party plugins built on non-default toolchains. Genuinely irreversible
once such plugins ship, so done before wide adoption.

Enum width pinning: every by-value C-ABI enum (PJ_primitive_type_t,
PJ_data_source_state_t, the message-level / message-box-type / message-box-
buttons enums, PJ_toolbox_message_level_t) gains a
`..._FORCE_INT32 = 0x7FFFFFFF` enumerator. The existing sizeof==4
static_asserts only protect this project's build; the enumerator is what
forces a stable 4-byte width into a plugin compiled with -fshort-enums,
where these enums (max value 0xFF) would otherwise shrink to 1 byte and
silently misalign every struct/param that embeds them. Adds the missing
PJ_message_box_buttons_t sentinel.

Fixed-width sizes: replaces size_t with uint64_t in every C-ABI struct and
function-pointer signature (plugin_data_api.h). size_t is what made the ABI
implicitly 64-bit-only; uint64_t makes the width explicit and platform-
independent.

Both are layout-identical on the supported 64-bit target: enum sizeof stays
4, size_t and uint64_t are the same width, no vtable offset or size moves.
All layout sentinels and 62/62 tests pass unchanged.


* fix(abi): match C++ host/SDK to the uint64_t C ABI (fixes macOS build)

The prior commit widened the C ABI (plugin_data_api.h) from size_t to
uint64_t, but the C++ host vtable implementations and SDK host-views that
fill those slots still used size_t. On macOS uint64_t is `unsigned long
long` while size_t is `unsigned long` — same width, distinct types — so the
trampolines fail to bind to the slots and the out-param pointers mismatch
(size_t* vs uint64_t*). Linux (glibc) and Windows (MSVC) alias both to one
type, so only macOS CI caught it.

Convert every ABI-boundary size_t to uint64_t, keeping size_t for purely
internal in-memory indexing:
  - pj_datastore host trampolines: append_record/append_bound_record
    field_count, push_owned size, lazy-fetch out_size, set_retention_budget,
    list_topics capacity/out_count, get_bytes out_size, entry_count return.
  - SDK host-views (plugin_data_api.hpp): list_topics/get_string_list counts,
    entryCount return, setRetentionBudget param, both LazyBox trampolines;
    object_bytes.hpp get_bytes; settings_store_host slots.
  - testing helpers + 4 built test trampolines bound to the same slots.

Value conversions size_t<->uint64_t are same-width and warning-free on
64-bit; only the pointer and function-pointer types required fixing.
Internal storage (RetentionBudget.max_memory_bytes, encoding/buffer
indexing) stays size_t.

Build (Debug+ASAN, -Werror -Wconversion) + 62/62 tests pass; clang-format clean.


---------

+++ pj_base/include/pj_base/sdk/settings_store_host.hpp
@@ -105,7 +105,7 @@ class SettingsStoreHost {
   }
 
   static bool tGetStringList(
-      void* ctx, PJ_string_view_t key, const PJ_string_view_t** out_items, size_t* out_count, bool* out_found,
+      void* ctx, PJ_string_view_t key, const PJ_string_view_t** out_items, uint64_t* out_count, bool* out_found,
       PJ_error_t* out_error) noexcept {
     if (ctx == nullptr || out_items == nullptr || out_count == nullptr || out_found == nullptr) {
       fillError(out_error, 2, "settings", "null ctx or out-param");
@@ -140,7 +140,7 @@ class SettingsStoreHost {
   }
 
   static bool tSetStringList(
-      void* ctx, PJ_string_view_t key, const PJ_string_view_t* items, size_t count, PJ_error_t* out_error) noexcept {
+      void* ctx, PJ_string_view_t key, const PJ_string_view_t* items, uint64_t count, PJ_error_t* out_error) noexcept {
     if (ctx == nullptr || (items == nullptr && count != 0)) {
       fillError(out_error, 2, "settings", "null ctx or items");
       return false;
